refactor agentic pipeline and tighten tool access
