+++ algorithm/chat/app/core/chat_server.py
@@ -4,7 +4,7 @@ from fastapi import FastAPI
 from lazyllm import LOG, once_wrapper
 
 from chat.config import URL_MAP, SENSITIVE_WORDS_PATH, DEFAULT_CHAT_DATASET
-from chat.pipelines.agentic_v2 import agentic_rag_v2
+from chat.pipelines.agentic import agentic_rag
 from chat.pipelines.naive import get_ppl_naive
 from chat.components.process.sensitive_filter import SensitiveFilter
 
@@ -35,7 +35,7 @@ class ChatServer:
         try:
             self.query_ppl: Dict[str, Any] = {}
             self.query_ppl_stream: Dict[str, Any] = {}
-            self.query_ppl_reasoning = agentic_rag_v2
+            self.query_ppl_reasoning = agentic_rag
             self.sensitive_filter = SensitiveFilter(SENSITIVE_WORDS_PATH)
 
             if self.sensitive_filter.loaded:
