+++ algorithm/chat/runtime_models.inner.yaml
@@ -72,7 +72,7 @@ retrieval:
   #     embed_keys: [bge_m3_dense]
   #     topk: 20
 
-agentic_v2:
+agentic:
   kb_url: ${LAZYRAG_AGENTIC_KB_URL:-http://lazyllm-algo:8000}
   core_api_url: ${LAZYRAG_CORE_API_URL:-http://core:8000}
   kb_name: ${LAZYRAG_AGENTIC_KB_NAME:-general_algo}
